refactored code for workSpaceTreeView
