+++ src/components/main/actionsPanel/workspaceTreeView/WorkspaceTreeView.tsx
@@ -4,76 +4,45 @@ import React, {
   useEffect,
   useMemo,
   useRef,
-  useState,
 } from "react";
-
 import cx from "classnames";
-import {
-  FileSystemDirectoryHandle,
-  FileSystemFileHandle,
-  FileSystemHandle,
-} from "file-system-access";
-import { DraggingPositionItem, TreeItem } from "react-complex-tree";
+import { DraggingPositionItem } from "react-complex-tree";
 import { useDispatch, useSelector } from "react-redux";
 
-import { SVGIconI, SVGIconII, TreeView } from "@_components/common";
+import { SVGIconI, TreeView } from "@_components/common";
 import { TreeViewData } from "@_components/common/treeView/types";
 import {
   AddFileActionPrefix,
-  DefaultProjectPath,
   HmsClearActionType,
-  LogAllow,
-  ParsableFileTypes,
   RootNodeUid,
-  TmpNodeUid,
 } from "@_constants/main";
-import { getValidNodeUids } from "@_node/apis";
 import {
   _path,
-  createDirectory,
   getNormalizedPath,
-  getStat,
-  readDir,
-  readFile,
-  reloadIDBProject,
-  reloadLocalProject,
-  removeFileSystem,
-  TFileHandlerCollection,
   TFileNodeData,
   TFilesReference,
-  writeFile,
 } from "@_node/file";
-import { TNode, TNodeTreeData, TNodeUid, TNormalNodeData } from "@_node/types";
+import { TNode, TNodeUid } from "@_node/types";
 import {
-  collapseFFNode,
-  expandFFNode,
   ffSelector,
-  focusFFNode,
   globalSelector,
   MainContext,
   navigatorSelector,
-  removeCurrentFile,
-  selectFFNode,
-  setCurrentFile,
-  setFileAction,
-  updateFFTreeViewState,
 } from "@_redux/main";
 import {
   addClass,
   generateQuerySelector,
   removeClass,
-  verifyFileHandlerPermission,
 } from "@_services/main";
-import { TFileAction, TFileNodeType } from "@_types/main";
 import { 
-  deletingWarning, 
-  duplicatingWarning, 
-  fileError, 
-  folderError, 
-  invalidDirError, 
-  movingError, 
-  renamingError 
-} from "./errors";
+  useFileOperations, 
+  useInvalidNodes, 
+  useNodeActionsHandler, 
+  useNodeViewState, 
+  useTemporaryNodes,
+  useCmdk
+} from "./hooks";
+import { Container, ItemArrow } from "./workSpaceTreeComponents";
 
 const AutoExpandDelay = 1 * 1000;
 export default function WorkspaceTreeView() {
@@ -84,9 +53,7 @@ export default function WorkspaceTreeView() {
   const {
     focusedItem,
     expandedItems,
-    expandedItemsObj,
     selectedItems,
-    selectedItemsObj,
   } = useSelector(ffSelector);
   const {
     // global action
@@ -95,250 +62,40 @@ export default function WorkspaceTreeView() {
     // navigator
     project,
     navigatorDropDownType,
-    setNavigatorDropDownType,
     // node actions
     activePanel,
     setActivePanel,
-    clipboardData,
-    setClipboardData,
     // actions panel
     showActionsPanel,
     // file tree view
     initialFileToOpen,
     setInitialFileToOpen,
     ffTree,
-    setFFTree,
-    ffHandlers,
-    setFFHandlers,
     ffHoveredItem,
     setFFHoveredItem,
     isHms,
     setIsHms,
     ffAction,
     currentFileUid,
-    setCurrentFileUid,
-    // node tree view
-    nodeTree,
-    setNodeTree,
-    setValidNodeTree,
     // stage view
-    setIFrameSrc,
     linkToOpen,
-    // code view
-    showCodeView,
-    setShowCodeView,
-    // processor
-    setUpdateOpt,
     // references
     filesReferenceData,
-    htmlReferenceData,
     // cmdk
     currentCommand,
-    // other
-    osType,
-    // toasts
-    addMessage,
     // non-parse file
     parseFileFlag,
     setParseFile,
     prevFileUid,
-    setPrevFileUid,
   } = useContext(MainContext);
   // -------------------------------------------------------------- node status --------------------------------------------------------------
   //  invalid - can't do any actions on the nodes
-  const [invalidNodes, _setInvalidNodes] = useState<{
-    [uid: TNodeUid]: boolean;
-  }>({});
-  const setInvalidNodes = useCallback(
-    (...uids: TNodeUid[]) => {
-      const _invalidNodes = { ...invalidNodes };
-      uids.map((uid) => (_invalidNodes[uid] = true));
-      _setInvalidNodes(_invalidNodes);
-    },
-    [invalidNodes],
-  );
-  const removeInvalidNodes = useCallback(
-    (...uids: TNodeUid[]) => {
-      const _invalidNodes = { ...invalidNodes };
-      uids.map((uid) => delete _invalidNodes[uid]);
-      _setInvalidNodes(_invalidNodes);
-    },
-    [invalidNodes],
-  );
+  const {invalidNodes} = useInvalidNodes()
   // temporary - don't display the nodes
-  const [temporaryNodes, _setTemporaryNodes] = useState<{
-    [uid: TNodeUid]: boolean;
-  }>({});
-  const setTemporaryNodes = useCallback(
-    (...uids: TNodeUid[]) => {
-      const _temporaryNodes = { ...temporaryNodes };
-      uids.map((uid) => (_temporaryNodes[uid] = true));
-      _setTemporaryNodes(_temporaryNodes);
-    },
-    [temporaryNodes],
-  );
-  const removeTemporaryNodes = useCallback(
-    (...uids: TNodeUid[]) => {
-      const _temporaryNodes = { ...temporaryNodes };
-      uids.map((uid) => delete _temporaryNodes[uid]);
-      _setTemporaryNodes(_temporaryNodes);
-    },
-    [temporaryNodes],
-  );
-  // -------------------------------------------------------------- project load cb for side effect --------------------------------------------------------------
-  const ffTreeRef = useRef<TNodeTreeData>({});
-  const cb_reloadProject = useCallback(
-    async (_uid?: TNodeUid) => {
-      const treeViewData: TNodeTreeData = {};
-      const ffHandlerObj: TFileHandlerCollection = {};
-      let _deletedUids: TNodeUid[] = [];
-
-      if (project.context === "local") {
-        try {
-          const { handlerObj, deletedUids } = await reloadLocalProject(
-            ffHandlers[RootNodeUid] as FileSystemDirectoryHandle,
-            ffTree,
-            osType,
-          );
-          _deletedUids = deletedUids;
-
-          // sort by ASC directory/file
-          Object.keys(handlerObj).map((uid) => {
-            const handler = handlerObj[uid];
-            handler.children = handler.children.sort((a, b) => {
-              return handlerObj[a].kind === "file" &&
-                handlerObj[b].kind === "directory"
-                ? 1
-                : handlerObj[a].kind === "directory" &&
-                  handlerObj[b].kind === "file"
-                ? -1
-                : handlerObj[a].name > handlerObj[b].name
-                ? 1
-                : -1;
-            });
-          });
-
-          // set ff-tree, ff-handlers
-          Object.keys(handlerObj).map((uid) => {
-            const {
-              parentUid,
-              children,
-              path,
-              kind,
-              name,
-              ext,
-              content,
-              handler,
-            } = handlerObj[uid];
-            const type = ParsableFileTypes[ext || ""]
-              ? ext?.slice(1)
-              : "unknown";
-            treeViewData[uid] = {
-              uid,
-              parentUid: parentUid,
-              name: name,
-              isEntity: kind === "file",
-              children: [...children],
-              data: {
-                valid: true,
-                path: path,
-                kind: kind,
-                name: name,
-                ext: ext,
-                type,
-                orgContent:
-                  type !== "unknown"
-                    ? ffTree[uid]
-                      ? (ffTree[uid].data as TFileNodeData).orgContent
-                      : content?.toString()
-                    : "",
-                content:
-                  type !== "unknown"
-                    ? ffTree[uid]
-                      ? (ffTree[uid].data as TFileNodeData).content
-                      : content?.toString()
-                    : "",
-                changed: ffTree[uid]
-                  ? (ffTree[uid].data as TFileNodeData).changed
-                  : false,
-              } as TFileNodeData,
-            } as TNode;
-
-            ffHandlerObj[uid] = handler;
-          });
-        } catch (err) {
-          LogAllow && console.log("failed to reload local project");
-        }
-      } else if (project.context === "idb") {
-        try {
-          const { handlerObj, deletedUids } = await reloadIDBProject(
-            DefaultProjectPath,
-            ffTree,
-          );
-          _deletedUids = deletedUids;
-
-          // sort by ASC directory/file
-          Object.keys(handlerObj).map((uid) => {
-            const handler = handlerObj[uid];
-            handler.children = handler.children.sort((a, b) => {
-              return handlerObj[a].kind === "file" &&
-                handlerObj[b].kind === "directory"
-                ? 1
-                : handlerObj[a].kind === "directory" &&
-                  handlerObj[b].kind === "file"
-                ? -1
-                : handlerObj[a].name > handlerObj[b].name
-                ? 1
-                : -1;
-            });
-          });
-
-          // set ff-tree, ff-handlers
-          Object.keys(handlerObj).map((uid) => {
-            const { parentUid, children, path, kind, name, ext, content } =
-              handlerObj[uid];
-            const type = ParsableFileTypes[ext || ""]
-              ? ext?.slice(1)
-              : "unknown";
-            treeViewData[uid] = {
-              uid,
-              parentUid: parentUid,
-              name: name,
-              isEntity: kind === "file",
-              children: [...children],
-              data: {
-                valid: true,
-                path: path,
-                kind: kind,
-                name: name,
-                ext: ext,
-                type,
-                orgContent: type !== "unknown" ? content?.toString() : "",
-                content: type !== "unknown" ? content?.toString() : "",
-                changed: false,
-              } as TFileNodeData,
-            } as TNode;
-          });
-        } catch (err) {
-          LogAllow && console.log("failed to reload welcome project");
-        }
-      }
-
-      dispatch(updateFFTreeViewState({ deletedUids: _deletedUids }));
-      if (_uid && !treeViewData[_uid]) {
-        setIFrameSrc(null);
-        setNodeTree({});
-        setValidNodeTree({});
-        setCurrentFileUid("");
-        dispatch(removeCurrentFile());
-      }
-      ffTreeRef.current = treeViewData;
-      setFFTree(treeViewData);
-      setFFHandlers(ffHandlerObj);
-    },
-    [project.context, ffHandlers, ffTree, osType],
-  );
+  const {setTemporaryNodes, removeTemporaryNodes} = useTemporaryNodes()
   // -------------------------------------------------------------- hms --------------------------------------------------------------
+	const { _copy, _create, _cut, _delete, _rename } = useFileOperations()
+  
   useEffect(() => {
     if (isHms === null) return;
 
@@ -417,319 +174,6 @@ export default function WorkspaceTreeView() {
 
     setIsHms(null);
   }, [isHms]);
-  const _create = useCallback(
-    async (params: {
-      parentUid: TNodeUid;
-      name: string;
-      type: TFileNodeType;
-    }) => {
-      addRunningActions(["fileTreeView-create"]);
-
-      const { parentUid, name, type } = params;
-      try {
-        // validate
-        const parentNode = ffTree[parentUid];
-        if (parentNode === undefined) throw "error";
-        const parentNodeData = parentNode.data as TFileNodeData;
-
-        if (project.context === "local") {
-          // verify handler permission
-          const parentHandler = ffHandlers[
-            parentNode.uid
-          ] as FileSystemDirectoryHandle;
-          if (!(await verifyFileHandlerPermission(parentHandler)))
-            throw "error";
-
-          // create
-          if (type === "*folder") {
-            await parentHandler.getDirectoryHandle(name, { create: true });
-          } else {
-            // file
-            await parentHandler.getFileHandle(name, { create: true });
-          }
-        } else if (project.context === "idb") {
-          // create
-          if (type === "*folder") {
-            await createDirectory(`${parentNodeData.path}/${name}`);
-          } else {
-            await writeFile(`${parentNodeData.path}/${name}`, "");
-          }
-        }
-      } catch (err) {}
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-create"], false);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      ffTree,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
-  const _delete = useCallback(
-    async (uids: TNodeUid[]) => {
-      addRunningActions(["fileTreeView-delete"]);
-      setInvalidNodes(...uids);
-
-      await Promise.all(
-        uids.map(async (uid) => {
-          try {
-            // validate
-            const node = ffTree[uid];
-            if (node === undefined) throw "error";
-            const nodeData = node.data as TFileNodeData;
-            const parentNode = ffTree[node.parentUid as TNodeUid];
-            if (parentNode === undefined) throw "error";
-            const parentNodeData = parentNode.data as TFileNodeData;
-
-            if (project.context === "local") {
-              const parentHandler = ffHandlers[
-                parentNode.uid
-              ] as FileSystemDirectoryHandle;
-              if (!(await verifyFileHandlerPermission(parentHandler)))
-                throw "error";
-
-              // delete
-              try {
-                const entryName =
-                  nodeData.kind === "directory"
-                    ? nodeData.name
-                    : `${nodeData.name}${nodeData.ext}`;
-                await parentHandler.removeEntry(entryName, { recursive: true });
-              } catch (err) {}
-            } else if (project.context === "idb") {
-              // delete
-              try {
-                const entryName =
-                  nodeData.kind === "directory"
-                    ? nodeData.name
-                    : `${nodeData.name}${nodeData.ext}`;
-                await removeFileSystem(`${parentNodeData.path}/${entryName}`);
-              } catch (err) {}
-            }
-          } catch (err) {}
-        }),
-      );
-
-      removeInvalidNodes(...uids);
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-delete"], false);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      setInvalidNodes,
-      removeInvalidNodes,
-      ffTree,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
-  const _rename = useCallback(
-    async (uid: TNodeUid, newName: string) => {
-      addRunningActions(["fileTreeView-rename"]);
-
-      try {
-        // validate
-        const node = ffTree[uid];
-        if (node === undefined || node.name === newName) throw "error";
-        const nodeData = node.data as TFileNodeData;
-        const parentNode = ffTree[node.parentUid as TNodeUid];
-        if (parentNode === undefined) throw "error";
-        const parentNodeData = parentNode.data as TFileNodeData;
-
-        const newUid = `${parentNode.uid}/${newName}`;
-        setTemporaryNodes(uid);
-        setInvalidNodes(newUid);
-
-        if (project.context === "local") {
-          const handler = ffHandlers[uid],
-            parentHandler = ffHandlers[
-              parentNode.uid
-            ] as FileSystemDirectoryHandle;
-          if (
-            !(await verifyFileHandlerPermission(handler)) ||
-            !(await verifyFileHandlerPermission(parentHandler))
-          )
-            throw "error";
-
-          await moveLocalFF(handler, parentHandler, parentHandler, newName);
-        } else if (project.context === "idb") {
-          await moveIDBFF(nodeData, parentNodeData, newName);
-        }
-
-        removeInvalidNodes(newUid);
-        removeTemporaryNodes(uid);
-      } catch (err) {}
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-rename"], false);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      setTemporaryNodes,
-      removeTemporaryNodes,
-      setInvalidNodes,
-      removeInvalidNodes,
-      ffTree,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
-  const _cut = useCallback(
-    async (uids: TNodeUid[], targetUids: TNodeUid[]) => {
-      addRunningActions(["fileTreeView-move"]);
-
-      const _invalidNodes = { ...invalidNodes };
-
-      await Promise.all(
-        uids.map(async (uid, index) => {
-          const targetUid = targetUids[index];
-
-          // validate
-          const node = ffTree[uid];
-          if (node === undefined) return;
-          const nodeData = node.data as TFileNodeData;
-          const parentNode = ffTree[node.parentUid as TNodeUid];
-          if (parentNode === undefined) return;
-          const targetNode = ffTree[targetUid];
-          if (targetNode === undefined) return;
-          const targetNodeData = targetNode.data as TFileNodeData;
-
-          const newUid = `${targetUid}/${nodeData.name}`;
-          _invalidNodes[uid] = true;
-          _invalidNodes[newUid] = true;
-          setInvalidNodes(...Object.keys(_invalidNodes));
-
-          if (project.context === "local") {
-            const handler = ffHandlers[uid],
-              parentHandler = ffHandlers[
-                parentNode.uid
-              ] as FileSystemDirectoryHandle,
-              targetHandler = ffHandlers[
-                targetUid
-              ] as FileSystemDirectoryHandle;
-            if (
-              !(await verifyFileHandlerPermission(handler)) ||
-              !(await verifyFileHandlerPermission(parentHandler)) ||
-              !(await verifyFileHandlerPermission(targetHandler))
-            )
-              return;
-
-            try {
-              await moveLocalFF(
-                handler,
-                parentHandler,
-                targetHandler,
-                nodeData.name,
-              );
-            } catch (err) {}
-          } else if (project.context === "idb") {
-            await moveIDBFF(nodeData, targetNodeData, nodeData.name);
-          }
-
-          delete _invalidNodes[uid];
-          delete _invalidNodes[newUid];
-          setInvalidNodes(...Object.keys(_invalidNodes));
-        }),
-      );
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-move"], false);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      invalidNodes,
-      setInvalidNodes,
-      ffTree,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
-  const _copy = useCallback(
-    async (uids: TNodeUid[], names: string[], targetUids: TNodeUid[]) => {
-      addRunningActions(["fileTreeView-duplicate"]);
-
-      const _invalidNodes = { ...invalidNodes };
-
-      await Promise.all(
-        uids.map(async (uid, index) => {
-          const name = names[index];
-          const targetUid = targetUids[index];
-
-          // validate
-          const node = ffTree[uid];
-          if (node === undefined) return;
-          const nodeData = node.data as TFileNodeData;
-          const parentNode = ffTree[node.parentUid as TNodeUid];
-          if (parentNode === undefined) return;
-          const targetNode = ffTree[targetUid];
-          if (targetNode === undefined) return;
-          const targetNodeData = targetNode.data as TFileNodeData;
-
-          const newUid = `${targetUid}/${name}`;
-          _invalidNodes[uid] = true;
-          _invalidNodes[newUid] = true;
-          setInvalidNodes(...Object.keys(_invalidNodes));
-
-          if (project.context === "local") {
-            const handler = ffHandlers[uid],
-              parentHandler = ffHandlers[
-                parentNode.uid
-              ] as FileSystemDirectoryHandle,
-              targetHandler = ffHandlers[
-                targetUid
-              ] as FileSystemDirectoryHandle;
-            if (
-              !(await verifyFileHandlerPermission(handler)) ||
-              !(await verifyFileHandlerPermission(parentHandler)) ||
-              !(await verifyFileHandlerPermission(targetHandler))
-            )
-              return;
-
-            try {
-              await moveLocalFF(
-                handler,
-                parentHandler,
-                targetHandler,
-                name,
-                true,
-              );
-            } catch (err) {
-              console.log(err);
-            }
-          } else if (project.context === "idb") {
-            await moveIDBFF(nodeData, targetNodeData, nodeData.name, true);
-          }
-
-          delete _invalidNodes[uid];
-          delete _invalidNodes[newUid];
-          setInvalidNodes(...Object.keys(_invalidNodes));
-        }),
-      );
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-duplicate"], false);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      invalidNodes,
-      setInvalidNodes,
-      ffTree,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
   // -------------------------------------------------------------- sync --------------------------------------------------------------
   // outline the hovered item
   const hoveredItemRef = useRef<TNodeUid>(ffHoveredItem);
@@ -790,138 +234,12 @@ export default function WorkspaceTreeView() {
     return data;
   }, [ffTree]);
   // -------------------------------------------------------------- node view state handlers --------------------------------------------------------------
-  const cb_focusNode = useCallback(
-    (uid: TNodeUid) => {
-      // validate
-      if (
-        invalidNodes[uid] ||
-        focusedItem === uid ||
-        ffTree[uid] === undefined
-      ) {
-        removeRunningActions(["fileTreeView-focus"], false);
-        return;
-      }
-
-      addRunningActions(["fileTreeView-focus"]);
-      dispatch(focusFFNode(uid));
-      removeRunningActions(["fileTreeView-focus"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      invalidNodes,
-      focusedItem,
-      ffTree,
-    ],
-  );
-  const cb_selectNode = useCallback(
-    (uids: TNodeUid[]) => {
-      // validate
-      let _uids = [...uids];
-      _uids = _uids.filter((_uid) => {
-        return !(ffTree[_uid] === undefined);
-      });
-      if (_uids.length === 0) {
-        removeRunningActions(["fileTreeView-select"], false);
-        return;
-      }
-      _uids = getValidNodeUids(ffTree, _uids);
-      if (_uids.length === selectedItems.length) {
-        let same = true;
-        for (const _uid of _uids) {
-          if (selectedItemsObj[_uid] === undefined) {
-            same = false;
-            break;
-          }
-        }
-        if (same) {
-          removeRunningActions(["fileTreeView-select"], false);
-          return;
-        }
-      }
-      if (project && file) {
-        // remove exist script
-        const exist = document.head.querySelector("#custom-plausible");
-        if (exist !== null) {
-          document.head.removeChild(exist);
-        }
-        // plausible analytics
-        var script = document.createElement("script");
-        script.id = "custom-plausible";
-        script.textContent =
-          `
-        plausible('pageview', { u: '` +
-          "rnbw.dev/" +
-          project.name +
-          "/" +
-          file.uid.replace("ROOT/", "") +
-          `' + window.location.search });
-      `;
-        document.head.appendChild(script);
-      }
-      addRunningActions(["fileTreeView-select"]);
-      dispatch(selectFFNode(_uids));
-      removeRunningActions(["fileTreeView-select"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      ffTree,
-      invalidNodes,
-      selectedItems,
-      selectedItemsObj,
-    ],
-  );
-  const cb_expandNode = useCallback(
-    (uid: TNodeUid) => {
-      // validate
-      if (
-        invalidNodes[uid] ||
-        ffTree[uid] === undefined ||
-        ffTree[uid].isEntity ||
-        expandedItemsObj[uid]
-      ) {
-        removeRunningActions(["fileTreeView-expand"], false);
-        return;
-      }
-
-      addRunningActions(["fileTreeView-expand"]);
-      dispatch(expandFFNode([uid]));
-      removeRunningActions(["fileTreeView-expand"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      invalidNodes,
-      ffTree,
-      expandedItemsObj,
-    ],
-  );
-  const cb_collapseNode = useCallback(
-    (uid: TNodeUid) => {
-      // validate
-      if (
-        invalidNodes[uid] ||
-        ffTree[uid] === undefined ||
-        ffTree[uid].isEntity ||
-        !expandedItemsObj[uid]
-      ) {
-        removeRunningActions(["fileTreeView-collapse"], false);
-        return;
-      }
-
-      addRunningActions(["fileTreeView-collapse"]);
-      dispatch(collapseFFNode([uid]));
-      removeRunningActions(["fileTreeView-collapse"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      invalidNodes,
-      ffTree,
-      expandedItemsObj,
-    ],
-  );
+  const {
+    cb_collapseNode,
+    cb_expandNode,
+    cb_focusNode,
+    cb_selectNode,
+  } = useNodeViewState()
   // -------------------------------------------------------------- project --------------------------------------------------------------
   // open default initial html file
   useEffect(() => {
@@ -959,446 +277,17 @@ export default function WorkspaceTreeView() {
     }
   }, [initialFileToOpen]);
   // -------------------------------------------------------------- node actions handlers --------------------------------------------------------------
-  const moveLocalFF = async (
-    handler: FileSystemHandle,
-    parentHandler: FileSystemDirectoryHandle,
-    targetHandler: FileSystemDirectoryHandle,
-    newName: string,
-    copy: boolean = false,
-    showWarning: boolean = false,
-  ) => {
-    if (handler.kind === "directory") {
-      // validate if the new name exists
-      let exists = true;
-      try {
-        await targetHandler.getDirectoryHandle(newName, { create: false });
-        exists = true;
-      } catch (err) {
-        exists = false;
-      }
-      if (exists) {
-        showWarning &&
-          addMessage({
-            type: "error",
-            content: "Folder with the same name already exists.",
-          });
-        return;
-      }
-
-      // move nested handler-dir to targetHandler with the newName - copy (optional)
-      try {
-        const newHandler = await targetHandler.getDirectoryHandle(newName, {
-          create: true,
-        });
-        const newDirHandlers = [newHandler];
-        const dirHandlers = [handler as FileSystemDirectoryHandle];
-        while (dirHandlers.length) {
-          const dirHandler = dirHandlers.shift() as FileSystemDirectoryHandle;
-          const newDirHandler =
-            newDirHandlers.shift() as FileSystemDirectoryHandle;
-          for await (const entry of dirHandler.values()) {
-            if (entry.kind === "directory") {
-              const newDir = await newDirHandler.getDirectoryHandle(
-                entry.name,
-                { create: true },
-              );
-              dirHandlers.push(entry);
-              newDirHandlers.push(newDir);
-            } else {
-              const newFile = await newDirHandler.getFileHandle(entry.name, {
-                create: true,
-              });
-              const content = await (entry as FileSystemFileHandle).getFile();
-              const writableStream = await newFile.createWritable();
-              await writableStream.write(content);
-              await writableStream.close();
-            }
-          }
-        }
-
-        // handle copy(optional)
-        !copy &&
-          (await parentHandler.removeEntry(handler.name, { recursive: true }));
-      } catch (err) {
-        throw "error";
-      }
-    } else {
-      // validate if the new name exists
-      let exists = true;
-      try {
-        await targetHandler.getFileHandle(newName, { create: false });
-        exists = true;
-      } catch (err) {
-        exists = false;
-      }
-      if (exists) {
-        showWarning &&
-          addMessage({
-            type: "error",
-            content: "File with the same name already exists.",
-          });
-        return;
-      }
-
-      // create a new file with the new name and write the content
-      try {
-        const newFile = await targetHandler.getFileHandle(newName, {
-          create: true,
-        });
-        const content = await (handler as FileSystemFileHandle).getFile();
-        const writableStream = await newFile.createWritable();
-        await writableStream.write(content);
-        await writableStream.close();
-
-        // handle copy(optional)
-        !copy &&
-          (await parentHandler.removeEntry(handler.name, { recursive: true }));
-      } catch (err) {
-        throw "error";
-      }
-    }
-  };
-  const moveIDBFF = async (
-    nodeData: TFileNodeData,
-    targetNodeData: TFileNodeData,
-    newName: string,
-    copy: boolean = false,
-    showWarning: boolean = false,
-  ) => {
-    if (nodeData.kind === "directory") {
-      // validate if the new name exists
-      let exists = true;
-      try {
-        await getStat(`${targetNodeData.path}/${newName}`);
-        exists = true;
-      } catch (err) {
-        exists = false;
-      }
-      if (exists) {
-        showWarning &&
-          addMessage({
-            type: "error",
-            content: "Folder with the same name already exists.",
-          });
-        return;
-      }
-
-      // move nested handler-dir to targetHandler with the newName - copy (optional)
-      try {
-        const dirs = [
-          {
-            orgPath: nodeData.path,
-            newPath: `${targetNodeData.path}/${newName}`,
-          },
-        ];
-        while (dirs.length) {
-          const { orgPath, newPath } = dirs.shift() as {
-            orgPath: string;
-            newPath: string;
-          };
-          await createDirectory(newPath);
-
-          const entries = await readDir(orgPath);
-          await Promise.all(
-            entries.map(async (entry) => {
-              const c_orgPath = `${orgPath}/${entry}`;
-              const c_newPath = `${newPath}/${entry}`;
-              const stats = await getStat(c_orgPath);
-              const c_kind = stats.type === "DIRECTORY" ? "directory" : "file";
-              if (c_kind === "directory") {
-                dirs.push({ orgPath: c_orgPath, newPath: c_newPath });
-              } else {
-                await writeFile(c_newPath, await readFile(c_orgPath));
-              }
-            }),
-          );
-        }
-
-        // handle copy(optional)
-        !copy && (await removeFileSystem(nodeData.path));
-      } catch (err) {
-        throw "error";
-      }
-    } else {
-      // validate if the new name exists
-      let exists = true;
-      try {
-        await getStat(`${targetNodeData.path}/${newName}`);
-        exists = true;
-      } catch (err) {
-        exists = false;
-      }
-      if (exists) {
-        showWarning &&
-          addMessage({
-            type: "error",
-            content: "File with the same name already exists.",
-          });
-        return;
-      }
-
-      // create a new file with the new name and write the content
-      try {
-        await writeFile(
-          `${targetNodeData.path}/${newName}`,
-          await readFile(nodeData.path),
-        );
-
-        // handle copy(optional)
-        !copy && (await removeFileSystem(nodeData.path));
-      } catch (err) {
-        throw "error";
-      }
-    }
-  };
-
-  const createFFNode = useCallback(
-    async (parentUid: TNodeUid, ffType: TFileNodeType, ffName: string) => {
-      addRunningActions(["fileTreeView-create"]);
-
-      let newName: string = "";
-
-      if (project.context === "local") {
-        // validate
-        const parentHandler = ffHandlers[
-          parentUid
-        ] as FileSystemDirectoryHandle;
-        if (!(await verifyFileHandlerPermission(parentHandler))) {
-          addMessage(invalidDirError);
-          removeRunningActions(["fileTreeView-create"], false);
-          return;
-        }
-
-        if (ffType === "*folder") {
-          // generate new folder name - ex: {aaa - copy}...
-          let folderName = ffName;
-          let exists = true;
-          try {
-            await parentHandler.getDirectoryHandle(ffName, { create: false });
-            exists = true;
-          } catch (err) {
-            exists = false;
-          }
-
-          if (exists) {
-            let index = 0;
-            while (exists) {
-              const _folderName = `${ffName} (${++index})`;
-              try {
-                await parentHandler.getDirectoryHandle(_folderName, {
-                  create: false,
-                });
-                exists = true;
-              } catch (err) {
-                folderName = _folderName;
-                exists = false;
-              }
-            }
-          }
-
-          newName = folderName;
-
-          // create the directory with generated name
-          try {
-            await parentHandler.getDirectoryHandle(folderName, {
-              create: true,
-            });
-          } catch (err) {
-            addMessage(folderError);
-            removeRunningActions(["fileTreeView-create"], false);
-            return;
-          }
-        } else {
-          // file
-          // generate new file name - ex: {aaa - copy}...
-          let fileName = `${ffName}.${ffType}`;
-          let exists = true;
-          try {
-            await parentHandler.getFileHandle(`${ffName}.${ffType}`, {
-              create: false,
-            });
-            exists = true;
-          } catch (err) {
-            exists = false;
-          }
-
-          if (exists) {
-            let index = 0;
-            while (exists) {
-              const _fileName = `${ffName} (${++index}).${ffType}`;
-              try {
-                await parentHandler.getFileHandle(_fileName, { create: false });
-                exists = true;
-              } catch (err) {
-                fileName = _fileName;
-                exists = false;
-              }
-            }
-          }
-
-          newName = fileName;
-
-          // create the file with generated name
-          try {
-            await parentHandler.getFileHandle(fileName, { create: true });
-          } catch (err) {
-            addMessage(fileError);
-            removeRunningActions(["fileTreeView-create"], false);
-            return;
-          }
-        }
-      } else if (project.context === "idb") {
-        const parentNode = ffTree[parentUid];
-        const parentNodeData = parentNode.data as TFileNodeData;
-        if (ffType === "*folder") {
-          // generate new folder name - ex: {aaa - copy}...
-          let folderName = "";
-          let exists = true;
-          try {
-            folderName = ffName;
-            await getStat(`${parentNodeData.path}/${folderName}`);
-            exists = true;
-          } catch (err) {
-            exists = false;
-          }
-          if (exists) {
-            let index = 0;
-            while (exists) {
-              try {
-                folderName = `${ffName} (${++index})`;
-                await getStat(`${parentNodeData.path}/${folderName}`);
-                exists = true;
-              } catch (err) {
-                exists = false;
-              }
-            }
-          }
-          newName = folderName;
-
-          // create the directory with generated name
-          try {
-            await createDirectory(`${parentNodeData.path}/${folderName}`);
-          } catch (err) {
-            addMessage(folderError);
-            removeRunningActions(["fileTreeView-create"], false);
-            return;
-          }
-        } else {
-          // file
-          // generate new file name - ex: {aaa - copy}...
-          let fileName = "";
-          let exists = true;
-          try {
-            fileName = `${ffName}.${ffType}`;
-            await getStat(`${parentNodeData.path}/${fileName}`);
-            exists = true;
-          } catch (err) {
-            exists = false;
-          }
-          if (exists) {
-            let index = 0;
-            while (exists) {
-              try {
-                fileName = `${ffName} (${++index}).${ffType}`;
-                await getStat(`${parentNodeData.path}/${fileName}`);
-                exists = true;
-              } catch (err) {
-                exists = false;
-              }
-            }
-          }
-          newName = fileName;
-
-          // create the file with generated name
-          try {
-            await writeFile(`${parentNodeData.path}/${fileName}`, "");
-          } catch (err) {
-            addMessage(fileError);
-            removeRunningActions(["fileTreeView-create"], false);
-            return;
-          }
-        }
-      }
-
-      const action: TFileAction = {
-        type: "create",
-        param1: `${parentUid}/${newName}`,
-        param2: { parentUid, name: newName, type: ffType },
-      };
-      dispatch(setFileAction(action));
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-create"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
-
+  
   const openFileUid = useRef<TNodeUid>("");
-  const createTmpFFNode = useCallback(
-    async (ffNodeType: TFileNodeType) => {
-      const tmpTree = JSON.parse(JSON.stringify(ffTree)) as TNodeTreeData;
 
-      // validate
-      let node = tmpTree[focusedItem];
-      if (node === undefined) return;
-      if (node.isEntity) {
-        node = tmpTree[node.parentUid as TNodeUid];
-      }
-
-      // expand the focusedItem
-      node.uid !== RootNodeUid &&
-        expandedItemsObj[node.uid] === undefined &&
-        dispatch(expandFFNode([node.uid]));
-
-      // add tmp node
-      const tmpNode: TNode = {
-        uid: `${node.uid}/${TmpNodeUid}`,
-        parentUid: node.uid,
-        name:
-          ffNodeType === "*folder"
-            ? "Untitled"
-            : ffNodeType === "html"
-            ? "Untitled"
-            : "Untitled",
-        isEntity: ffNodeType !== "*folder",
-        children: [],
-        data: {
-          valid: false,
-          type: ffNodeType,
-        },
-      };
-
-      node.children.unshift(tmpNode.uid);
-      tmpTree[tmpNode.uid] = tmpNode;
-      // setFFTree(tmpTree)
-
-      setInvalidNodes(tmpNode.uid);
-      await createFFNode(node.uid as TNodeUid, tmpNode.data.type, tmpNode.name);
-      removeInvalidNodes(tmpNode.uid);
-      setNavigatorDropDownType("project");
-
-      if (ffNodeType !== "*folder") {
-        openFileUid.current = `${node.uid}/${tmpNode.name}.${ffNodeType}`;
-        setCurrentFileUid(openFileUid.current);
-      }
-    },
-    [
-      ffTree,
-      focusedItem,
-      expandedItemsObj,
-      setInvalidNodes,
-      createFFNode,
-      setNavigatorDropDownType,
-      removeInvalidNodes,
-      setFFTree,
-    ],
-  );
+  const {
+		cb_startRenamingNode,
+		cb_abortRenamingNode,
+		cb_renameNode,
+		cb_moveNode,
+		cb_readNode
+  } = useNodeActionsHandler(openFileUid)
+  
   useEffect(() => {
     if (
       ffTree[openFileUid.current] !== undefined &&
@@ -1407,933 +296,6 @@ export default function WorkspaceTreeView() {
       openFile(openFileUid.current);
     }
   }, [ffTree, currentFileUid]);
-  const cb_startRenamingNode = useCallback(
-    (uid: TNodeUid) => {
-      // validate
-      if (invalidNodes[uid]) {
-        removeInvalidNodes(uid);
-        return;
-      }
-      setInvalidNodes(uid);
-    },
-    [invalidNodes, setInvalidNodes, removeInvalidNodes],
-  );
-  const cb_abortRenamingNode = useCallback(
-    (item: TreeItem) => {
-      const node = item.data as TNode;
-      const nodeData = node.data as TFileNodeData;
-      if (!nodeData.valid) {
-        const tmpTree = JSON.parse(JSON.stringify(ffTree));
-        tmpTree[node.parentUid as TNodeUid].children = tmpTree[
-          node.parentUid as TNodeUid
-        ].children.filter((c_uid: TNodeUid) => c_uid !== node.uid);
-        delete tmpTree[item.data.uid];
-        setFFTree(tmpTree);
-      }
-      removeInvalidNodes(node.uid);
-    },
-    [ffTree, removeInvalidNodes],
-  );
-  const _cb_renameNode = useCallback(
-    async (uid: TNodeUid, newName: string, ext: string) => {
-      // validate
-      const node = ffTree[uid];
-      if (node === undefined || node.name === newName) return;
-      const nodeData = node.data as TFileNodeData;
-      const parentNode = ffTree[node.parentUid as TNodeUid];
-      if (parentNode === undefined) return;
-      const parentNodeData = parentNode.data as TFileNodeData;
-
-      addRunningActions(["fileTreeView-rename"]);
-
-      const _orgName =
-        ext === "*folder"
-          ? `${nodeData.name}`
-          : `${nodeData.name}${nodeData.ext}`;
-      const _newName = ext === "*folder" ? `${newName}` : `${newName}${ext}`;
-      const newUid = `${parentNode.uid}/${_newName}`;
-
-      if (project.context === "local") {
-        // verify handler permission
-        const handler = ffHandlers[uid],
-          parentHandler = ffHandlers[
-            parentNode.uid
-          ] as FileSystemDirectoryHandle;
-        if (
-          !(await verifyFileHandlerPermission(handler)) ||
-          !(await verifyFileHandlerPermission(parentHandler))
-        ) {
-          addMessage({
-            type: "error",
-            content: `Invalid directory/file. Check if you have "write" permission for the directory/file.`,
-          });
-
-          removeRunningActions(["fileTreeView-rename"], false);
-          return;
-        }
-
-        setInvalidNodes(newUid);
-
-        try {
-          await moveLocalFF(
-            handler,
-            parentHandler,
-            parentHandler,
-            _newName,
-            false,
-            true,
-          );
-          removeInvalidNodes(newUid);
-        } catch (err) {
-          addMessage(renamingError);
-
-          removeInvalidNodes(newUid);
-          removeRunningActions(["fileTreeView-rename"], false);
-          return;
-        }
-      } else if (project.context === "idb") {
-        setInvalidNodes(newUid);
-
-        try {
-          await moveIDBFF(nodeData, parentNodeData, _newName, false, true);
-          removeInvalidNodes(newUid);
-        } catch (err) {
-          addMessage(renamingError);
-
-          removeInvalidNodes(newUid);
-          removeRunningActions(["fileTreeView-rename"], false);
-          return;
-        }
-      }
-
-      const action: TFileAction = {
-        type: "rename",
-        param1: { uid, parentUid: parentNode.uid },
-        param2: { orgName: _orgName, newName: _newName },
-      };
-      dispatch(setFileAction(action));
-
-      // update redux
-      dispatch(
-        setCurrentFile({
-          uid: newUid,
-          parentUid: parentNode.uid,
-          name: nodeData.name,
-          content: nodeData.contentInApp as string,
-        }),
-      );
-      dispatch(updateFFTreeViewState({ convertedUids: [[uid, newUid]] }));
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-rename"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      setInvalidNodes,
-      removeInvalidNodes,
-      ffTree,
-      ffHandlers,
-      cb_reloadProject,
-    ],
-  );
-  const cb_renameNode = useCallback(
-    async (item: TreeItem, newName: string) => {
-      const node = item.data as TNode;
-      const nodeData = node.data as TNormalNodeData;
-      if (!invalidNodes[node.uid]) return;
-
-      if (nodeData.valid) {
-        const _file = ffTree[node.uid];
-        const _fileData = _file.data as TFileNodeData;
-
-        if (_file && _fileData.changed) {
-          // confirm
-          const message = `Your changes will be lost if you don't save them. Are you sure you want to continue without saving?`;
-          if (!window.confirm(message)) {
-            removeInvalidNodes(node.uid);
-            return;
-          }
-        }
-
-        setTemporaryNodes(_file.uid);
-        await _cb_renameNode(
-          _file.uid,
-          newName,
-          _fileData.kind === "directory" ? "*folder" : _fileData.ext,
-        );
-        removeTemporaryNodes(_file.uid);
-      } else {
-        await createFFNode(node.parentUid as TNodeUid, nodeData.type, newName);
-      }
-      removeInvalidNodes(node.uid);
-    },
-    [
-      invalidNodes,
-      _cb_renameNode,
-      setTemporaryNodes,
-      removeTemporaryNodes,
-      ffTree,
-      ffHandlers,
-      osType,
-      createFFNode,
-      removeInvalidNodes,
-    ],
-  );
-
-  const cb_deleteNode = useCallback(async () => {
-    // validate
-    const uids = selectedItems.filter((uid) => !invalidNodes[uid]);
-    if (uids.length === 0) return;
-
-    // confirm msgbox
-    const message = `Are you sure you want to delete them? This action cannot be undone!`;
-    if (!window.confirm(message)) return;
-
-    addRunningActions(["fileTreeView-delete"]);
-    setInvalidNodes(...uids);
-
-    if (project.context === "local") {
-      let allDone = true;
-      await Promise.all(
-        uids.map(async (uid) => {
-          // validate
-          const node = ffTree[uid];
-          if (node === undefined) {
-            allDone = false;
-            return;
-          }
-          const nodeData = node.data as TFileNodeData;
-          const parentNode = ffTree[node.parentUid as TNodeUid];
-          if (parentNode === undefined) {
-            allDone = false;
-            return;
-          }
-          const parentHandler = ffHandlers[
-            parentNode.uid
-          ] as FileSystemDirectoryHandle;
-          if (!(await verifyFileHandlerPermission(parentHandler))) {
-            allDone = false;
-            return;
-          }
-
-          // delete
-          try {
-            const entryName =
-              nodeData.kind === "directory"
-                ? nodeData.name
-                : `${nodeData.name}${nodeData.ext}`;
-            await parentHandler.removeEntry(entryName, { recursive: true });
-          } catch (err) {
-            allDone = false;
-          }
-        }),
-      );
-      if (!allDone) {
-        addMessage(deletingWarning);
-      }
-    } else if (project.context === "idb") {
-      let allDone = true;
-      await Promise.all(
-        uids.map(async (uid) => {
-          // validate
-          const node = ffTree[uid];
-          if (node === undefined) {
-            allDone = false;
-            return;
-          }
-          const nodeData = node.data as TFileNodeData;
-
-          // delete
-          try {
-            await removeFileSystem(nodeData.path);
-          } catch (err) {
-            allDone = false;
-          }
-        }),
-      );
-      if (!allDone) {
-        addMessage(deletingWarning);
-      }
-    }
-
-    removeInvalidNodes(...uids);
-    await cb_reloadProject(file.uid);
-    removeRunningActions(["fileTreeView-delete"], false);
-  }, [
-    addRunningActions,
-    removeRunningActions,
-    project.context,
-    invalidNodes,
-    setInvalidNodes,
-    removeInvalidNodes,
-    selectedItems,
-    ffTree,
-    ffHandlers,
-    cb_reloadProject,
-    file.uid,
-  ]);
-  const cb_moveNode = useCallback(
-    async (uids: TNodeUid[], targetUid: TNodeUid, copy: boolean = false) => {
-      // validate
-      const targetNode = ffTree[targetUid];
-      if (targetNode === undefined) return;
-      const validatedUids = getValidNodeUids(ffTree, uids, targetUid);
-      if (validatedUids.length === 0) return;
-
-      // confirm files' changes
-      let hasChangedFile = false;
-      validatedUids.map((uid) => {
-        const _file = ffTree[uid];
-        const _fileData = _file.data as TFileNodeData;
-        if (_file && _fileData.changed) {
-          hasChangedFile = true;
-        }
-      });
-      if (hasChangedFile) {
-        const message = `Your changes will be lost if you don't save them. Are you sure you want to continue without saving?`;
-        if (!window.confirm(message)) {
-          return;
-        }
-      }
-
-      addRunningActions(["fileTreeView-move"]);
-
-      const _uids: { uid: TNodeUid; parentUid: TNodeUid; name: string }[] = [];
-      const _invalidNodes = { ...invalidNodes };
-
-      if (project.context === "local") {
-        // verify target handler permission
-        const targetHandler = ffHandlers[
-          targetUid
-        ] as FileSystemDirectoryHandle;
-        if (!(await verifyFileHandlerPermission(targetHandler))) {
-          addMessage(invalidDirError);
-          removeRunningActions(["fileTreeView-move"], false);
-          return;
-        }
-
-        let allDone = true;
-        await Promise.all(
-          validatedUids.map(async (uid) => {
-            // validate
-            const node = ffTree[uid];
-            if (node === undefined) {
-              allDone = false;
-              return;
-            }
-            const nodeData = node.data as TFileNodeData;
-            const parentNode = ffTree[node.parentUid as TNodeUid];
-            if (parentNode === undefined) {
-              allDone = false;
-              return;
-            }
-            const handler = ffHandlers[uid],
-              parentHandler = ffHandlers[
-                parentNode.uid
-              ] as FileSystemDirectoryHandle;
-            if (
-              !(await verifyFileHandlerPermission(handler)) ||
-              !(await verifyFileHandlerPermission(parentHandler))
-            ) {
-              allDone = false;
-              return;
-            }
-
-            // generate new name
-            let newName =
-              nodeData.kind === "directory"
-                ? nodeData.name
-                : `${nodeData.name}${nodeData.ext}`;
-            if (copy) {
-              if (nodeData.kind === "directory") {
-                let folderName = "";
-                let exists = false;
-                try {
-                  folderName = nodeData.name;
-                  await targetHandler.getDirectoryHandle(folderName, {
-                    create: false,
-                  });
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-                if (exists) {
-                  try {
-                    folderName = `${nodeData.name} copy`;
-                    await targetHandler.getDirectoryHandle(folderName, {
-                      create: false,
-                    });
-                    exists = true;
-                  } catch (err) {
-                    exists = false;
-                  }
-                  if (exists) {
-                    let index = 0;
-                    while (exists) {
-                      try {
-                        folderName = `${nodeData.name} copy (${++index})`;
-                        await targetHandler.getDirectoryHandle(folderName, {
-                          create: false,
-                        });
-                        exists = true;
-                      } catch (err) {
-                        exists = false;
-                      }
-                    }
-                  }
-                }
-                newName = folderName;
-              } else {
-                let fileName = "";
-                let exists = false;
-                try {
-                  fileName = `${nodeData.name}${nodeData.ext}`;
-                  await targetHandler.getFileHandle(fileName, {
-                    create: false,
-                  });
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-                if (exists) {
-                  try {
-                    fileName = `${nodeData.name} copy${nodeData.ext}`;
-                    await targetHandler.getFileHandle(fileName, {
-                      create: false,
-                    });
-                    exists = true;
-                  } catch (err) {
-                    exists = false;
-                  }
-                  if (exists) {
-                    let index = 0;
-                    while (exists) {
-                      try {
-                        fileName = `${nodeData.name} copy (${++index})${
-                          nodeData.ext
-                        }`;
-                        await targetHandler.getFileHandle(fileName, {
-                          create: false,
-                        });
-                        exists = true;
-                      } catch (err) {
-                        exists = false;
-                      }
-                    }
-                  }
-                  newName = fileName;
-                }
-              }
-            }
-
-            // update invalidNodes
-            const newUid = `${targetUid}/${newName}`;
-            _invalidNodes[uid] = true;
-            _invalidNodes[newUid] = true;
-            setInvalidNodes(...Object.keys(_invalidNodes));
-
-            // move
-            try {
-              await moveLocalFF(
-                handler,
-                parentHandler,
-                targetHandler,
-                newName,
-                copy,
-              );
-              _uids.push({ uid, parentUid: parentNode.uid, name: newName });
-            } catch (err) {
-              allDone = false;
-            }
-
-            // update invalidNodes
-            delete _invalidNodes[uid];
-            delete _invalidNodes[newUid];
-            setInvalidNodes(...Object.keys(_invalidNodes));
-          }),
-        );
-        if (!allDone) {
-          addMessage(movingError);
-        }
-      } else if (project.context === "idb") {
-        const targetNodeData = targetNode.data as TFileNodeData;
-
-        let allDone = true;
-        await Promise.all(
-          validatedUids.map(async (uid) => {
-            // validate
-            const node = ffTree[uid];
-            if (node === undefined) {
-              allDone = false;
-              return;
-            }
-            const nodeData = node.data as TFileNodeData;
-
-            // generate new name
-            let newName =
-              nodeData.kind === "directory"
-                ? nodeData.name
-                : `${nodeData.name}${nodeData.ext}`;
-            if (copy) {
-              if (nodeData.kind === "directory") {
-                let folderName = "";
-                let exists = false;
-                try {
-                  folderName = nodeData.name;
-                  await getStat(`${targetNodeData.path}/${folderName}`);
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-                if (exists) {
-                  try {
-                    folderName = `${nodeData.name} copy`;
-                    await getStat(`${targetNodeData.path}/${folderName}`);
-                    exists = true;
-                  } catch (err) {
-                    exists = false;
-                  }
-                  if (exists) {
-                    let index = 0;
-                    while (exists) {
-                      try {
-                        folderName = `${nodeData.name} copy (${++index})`;
-                        await getStat(`${targetNodeData.path}/${folderName}`);
-                        exists = true;
-                      } catch (err) {
-                        exists = false;
-                      }
-                    }
-                  }
-                }
-                newName = folderName;
-              } else {
-                let fileName = "";
-                let exists = false;
-                try {
-                  fileName = `${nodeData.name}${nodeData.ext}`;
-                  await getStat(`${targetNodeData.path}/${fileName}`);
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-                if (exists) {
-                  try {
-                    fileName = `${nodeData.name} copy${nodeData.ext}`;
-                    await getStat(`${targetNodeData.path}/${fileName}`);
-                    exists = true;
-                  } catch (err) {
-                    exists = false;
-                  }
-                  if (exists) {
-                    let index = 0;
-                    while (exists) {
-                      try {
-                        fileName = `${nodeData.name} copy (${++index})${
-                          nodeData.ext
-                        }`;
-                        await getStat(`${targetNodeData.path}/${fileName}`);
-                        exists = true;
-                      } catch (err) {
-                        exists = false;
-                      }
-                    }
-                  }
-                }
-                newName = fileName;
-              }
-            }
-
-            // update invalidNodes
-            const newUid = `${targetUid}/${newName}`;
-            _invalidNodes[uid] = true;
-            _invalidNodes[newUid] = true;
-            setInvalidNodes(...Object.keys(_invalidNodes));
-
-            // move
-            try {
-              await moveIDBFF(nodeData, targetNodeData, newName, copy);
-              _uids.push({
-                uid,
-                parentUid: node.parentUid as TNodeUid,
-                name: newName,
-              });
-            } catch (err) {
-              allDone = false;
-            }
-
-            // update invalidNodes
-            delete _invalidNodes[uid];
-            delete _invalidNodes[newUid];
-            setInvalidNodes(...Object.keys(_invalidNodes));
-          }),
-        );
-        if (!allDone) {
-          addMessage(movingError);
-        }
-      }
-
-      const action: TFileAction = {
-        type: copy ? "copy" : "cut",
-        param1: _uids,
-        param2: _uids.map(() => targetUid),
-      };
-      dispatch(setFileAction(action));
-
-      await cb_reloadProject();
-      removeRunningActions(["fileTreeView-move"]);
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      project.context,
-      invalidNodes,
-      setInvalidNodes,
-      ffTree,
-      ffHandlers,
-      ,
-      cb_reloadProject,
-    ],
-  );
-  const cb_duplicateNode = useCallback(async () => {
-    // validate
-    const uids = selectedItems.filter((uid) => !invalidNodes[uid]);
-    if (uids.length === 0) return;
-
-    // confirm files' changes
-    let hasChangedFile = false;
-    uids.map((uid) => {
-      const _file = ffTree[uid];
-      const _fileData = _file.data as TFileNodeData;
-      if (_file && _fileData.changed) {
-        hasChangedFile = true;
-      }
-    });
-    if (hasChangedFile) {
-      const message = `Your changes will be lost if you don't save them. Are you sure you want to continue without saving?`;
-      if (!window.confirm(message)) {
-        return;
-      }
-    }
-
-    addRunningActions(["fileTreeView-duplicate"]);
-
-    const _uids: { uid: TNodeUid; name: string }[] = [];
-    const _targetUids: TNodeUid[] = [];
-    const _invalidNodes = { ...invalidNodes };
-
-    if (project.context === "local") {
-      let allDone = true;
-      await Promise.all(
-        uids.map(async (uid) => {
-          // validate
-          const node = ffTree[uid];
-          if (node === undefined) {
-            allDone = false;
-            return;
-          }
-          const nodeData = node.data as TFileNodeData;
-          const parentNode = ffTree[node.parentUid as TNodeUid];
-          if (parentNode === undefined) {
-            allDone = false;
-            return;
-          }
-          const handler = ffHandlers[node.uid],
-            parentHandler = ffHandlers[
-              parentNode.uid
-            ] as FileSystemDirectoryHandle;
-          if (
-            !(await verifyFileHandlerPermission(handler)) ||
-            !(await verifyFileHandlerPermission(parentHandler))
-          ) {
-            allDone = false;
-            return;
-          }
-
-          // generate new name
-          let newName =
-            nodeData.kind === "directory"
-              ? `${nodeData.name} copy`
-              : `${nodeData.name} copy${nodeData.ext}`;
-          if (nodeData.kind === "directory") {
-            let folderName = "";
-            let exists = false;
-            try {
-              folderName = `${nodeData.name} copy`;
-              await parentHandler.getDirectoryHandle(folderName, {
-                create: false,
-              });
-              exists = true;
-            } catch (err) {
-              exists = false;
-            }
-            if (exists) {
-              let index = 0;
-              while (exists) {
-                try {
-                  folderName = `${nodeData.name} copy (${++index})`;
-                  await parentHandler.getDirectoryHandle(folderName, {
-                    create: false,
-                  });
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-              }
-            }
-            newName = folderName;
-          } else {
-            let fileName = "";
-            let exists = true;
-            try {
-              fileName = `${nodeData.name} copy${nodeData.ext}`;
-              await parentHandler.getFileHandle(fileName, { create: false });
-              exists = true;
-            } catch (err) {
-              exists = false;
-            }
-            if (exists) {
-              let index = 0;
-              while (exists) {
-                try {
-                  fileName = `${nodeData.name} copy (${++index})${
-                    nodeData.ext
-                  }`;
-                  await parentHandler.getFileHandle(fileName, {
-                    create: false,
-                  });
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-              }
-            }
-            newName = fileName;
-          }
-
-          // update invalidNodes
-          const newUid = `${node.parentUid}/${newName}`;
-          _invalidNodes[uid] = true;
-          _invalidNodes[newUid] = true;
-          setInvalidNodes(...Object.keys(_invalidNodes));
-
-          // duplicate
-          try {
-            await moveLocalFF(
-              handler,
-              parentHandler,
-              parentHandler,
-              newName,
-              true,
-            );
-            _uids.push({ uid, name: newName });
-            _targetUids.push(parentNode.uid);
-          } catch (err) {
-            allDone = false;
-          }
-
-          // set invalid nodes
-          delete _invalidNodes[uid];
-          delete _invalidNodes[newUid];
-          setInvalidNodes(...Object.keys(_invalidNodes));
-        }),
-      );
-      if (!allDone) {
-        addMessage(duplicatingWarning);
-      }
-    } else if (project.context === "idb") {
-      let allDone = true;
-      await Promise.all(
-        uids.map(async (uid) => {
-          // validate
-          const node = ffTree[uid];
-          if (node === undefined) {
-            allDone = false;
-            return;
-          }
-          const nodeData = node.data as TFileNodeData;
-          const parentNode = ffTree[node.parentUid as TNodeUid];
-          if (parentNode === undefined) {
-            allDone = false;
-            return;
-          }
-          const parentNodeData = parentNode.data as TFileNodeData;
-
-          // generate new name
-          let newName =
-            nodeData.kind === "directory"
-              ? `${nodeData.name} copy`
-              : `${nodeData.name} copy${nodeData.ext}`;
-          if (nodeData.kind === "directory") {
-            let folderName = "";
-            let exists = false;
-            try {
-              folderName = `${nodeData.name} copy`;
-              await getStat(`${parentNodeData.path}/${folderName}`);
-              exists = true;
-            } catch (err) {
-              exists = false;
-            }
-            if (exists) {
-              let index = 0;
-              while (exists) {
-                try {
-                  folderName = `${nodeData.name} copy (${++index})`;
-                  await getStat(`${parentNodeData.path}/${folderName}`);
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-              }
-            }
-            newName = folderName;
-          } else {
-            let fileName = "";
-            let exists = false;
-            try {
-              fileName = `${nodeData.name} copy${nodeData.ext}`;
-              await getStat(`${parentNodeData.path}/${fileName}`);
-              exists = true;
-            } catch (err) {
-              exists = false;
-            }
-            if (exists) {
-              let index = 0;
-              while (exists) {
-                try {
-                  fileName = `${nodeData.name} copy (${++index})${
-                    nodeData.ext
-                  }`;
-                  await getStat(`${parentNodeData.path}/${fileName}`);
-                  exists = true;
-                } catch (err) {
-                  exists = false;
-                }
-              }
-            }
-            newName = fileName;
-          }
-
-          // update invalidNodes
-          const newUid = `${node.parentUid}/${newName}`;
-          _invalidNodes[uid] = true;
-          _invalidNodes[newUid] = true;
-          setInvalidNodes(...Object.keys(_invalidNodes));
-
-          // duplicate
-          try {
-            await moveIDBFF(nodeData, parentNodeData, newName, true);
-            _uids.push({ uid, name: newName });
-            _targetUids.push(parentNode.uid);
-          } catch (err) {
-            allDone = false;
-          }
-
-          // set invalid nodes
-          delete _invalidNodes[uid];
-          delete _invalidNodes[newUid];
-          setInvalidNodes(...Object.keys(_invalidNodes));
-        }),
-      );
-      if (!allDone) {
-        addMessage(duplicatingWarning);
-      }
-    }
-
-    const action: TFileAction = {
-      type: "copy",
-      param1: _uids,
-      param2: _targetUids,
-    };
-    dispatch(setFileAction(action));
-
-    await cb_reloadProject();
-    removeRunningActions(["fileTreeView-duplicate"]);
-  }, [
-    addRunningActions,
-    removeRunningActions,
-    project.context,
-    invalidNodes,
-    setInvalidNodes,
-    selectedItems,
-    ffTree,
-    ffHandlers,
-    cb_reloadProject,
-  ]);
-
-  const cb_readNode = useCallback(
-    (uid: TNodeUid) => {
-      addRunningActions(["fileTreeView-read"]);
-      dispatch({ type: HmsClearActionType });
-      // validate
-      if (invalidNodes[uid]) {
-        removeRunningActions(["fileTreeView-read"], false);
-        return;
-      }
-      const node = ffTree[uid];
-      if (node === undefined || !node.isEntity || file.uid === uid) {
-        removeRunningActions(["fileTreeView-read"], false);
-        return;
-      }
-      const nodeData = node.data as TFileNodeData;
-      if (nodeData.type === "html") {
-        setPrevFileUid(file.uid);
-      }
-      if (nodeData.type === "unknown") {
-        dispatch(
-          setCurrentFile({
-            uid,
-            parentUid: node.parentUid as TNodeUid,
-            name: nodeData.name,
-            content: nodeData.content,
-          }),
-        );
-        removeRunningActions(["fileTreeView-read"]);
-        setParseFile(false);
-        showCodeView === false && setShowCodeView(true);
-      } else {
-        // set initial content of the html
-        let initialContent = "";
-        if (
-          nodeData.type === "html" &&
-          nodeData.kind === "file" &&
-          nodeData.content === ""
-        ) {
-          let doctype = "<!DOCTYPE html>\n";
-          let html = htmlReferenceData["elements"]["html"].Content
-            ? `<html>\n` +
-              htmlReferenceData["elements"]["html"].Content +
-              `\n</html>`
-            : "";
-          initialContent = doctype + html;
-          nodeData.content = initialContent;
-        }
-        addRunningActions(["processor-updateOpt"]);
-        dispatch(
-          setCurrentFile({
-            uid,
-            parentUid: node.parentUid as TNodeUid,
-            name: nodeData.name,
-            content: nodeData.content,
-          }),
-        );
-        setUpdateOpt({ parse: true, from: "file" });
-        setParseFile(true);
-        removeRunningActions(["fileTreeView-read"]);
-        setPrevFileUid(uid);
-      }
-    },
-    [
-      addRunningActions,
-      removeRunningActions,
-      invalidNodes,
-      ffTree,
-      file.uid,
-      showCodeView,
-    ],
-  );
 
   // handlle links-open
   const openFile = useCallback(
@@ -2369,6 +331,15 @@ export default function WorkspaceTreeView() {
     }
   }, [linkToOpen]);
   // -------------------------------------------------------------- cmdk --------------------------------------------------------------
+  const {
+    onDelete,
+		onCut,
+		onCopy,
+		onPaste,
+		onDuplicate,
+		onAddNode
+  } = useCmdk(openFileUid)
+  
   useEffect(() => {
     if (isAddFileAction(currentCommand.action)) {
       onAddNode(currentCommand.action);
@@ -2397,69 +368,11 @@ export default function WorkspaceTreeView() {
         break;
     }
   }, [currentCommand]);
-  const onDelete = useCallback(() => {
-    cb_deleteNode();
-  }, [cb_deleteNode]);
-  const onCut = useCallback(() => {
-    setClipboardData({
-      panel: "file",
-      type: "cut",
-      uids: selectedItems,
-      fileType: ffTree[file.uid].data.type,
-      data: [],
-      fileUid: file.uid,
-      prevNodeTree: nodeTree,
-    });
-  }, [selectedItems, ffTree[file.uid], nodeTree]);
-  const onCopy = useCallback(() => {
-    setClipboardData({
-      panel: "file",
-      type: "copy",
-      uids: selectedItems,
-      fileType: ffTree[file.uid].data.type,
-      data: [],
-      fileUid: file.uid,
-      prevNodeTree: nodeTree,
-    });
-  }, [selectedItems, ffTree[file.uid], nodeTree]);
-  const onPaste = useCallback(() => {
-    if (clipboardData.panel !== "file") return;
 
-    // validate
-    if (invalidNodes[focusedItem]) return;
-    const uids = clipboardData.uids.filter((uid) => !invalidNodes[uid]);
-    if (uids.length === 0) return;
-
-    if (clipboardData.type === "cut") {
-      setClipboardData({
-        panel: "file",
-        type: "cut",
-        uids: [],
-        fileType: "html",
-        data: [],
-        fileUid: "",
-        prevNodeTree: {},
-      });
-      cb_moveNode(uids, focusedItem);
-    } else if (clipboardData.type === "copy") {
-      cb_moveNode(uids, focusedItem, true);
-    }
-  }, [clipboardData, invalidNodes, focusedItem, cb_moveNode]);
-  const onDuplicate = useCallback(() => {
-    cb_duplicateNode();
-  }, [cb_duplicateNode]);
   const isAddFileAction = (actionName: string): boolean => {
     return actionName.startsWith(AddFileActionPrefix) ? true : false;
   };
-  const onAddNode = useCallback(
-    (actionName: string) => {
-      const nodeType = actionName.slice(AddFileActionPrefix.length + 1);
-      createTmpFFNode(
-        nodeType === "folder" ? "*folder" : (nodeType as TFileNodeType),
-      );
-    },
-    [createTmpFFNode],
-  );
+ 
   // -------------------------------------------------------------- own --------------------------------------------------------------
   const onPanelClick = useCallback((e: React.MouseEvent) => {
     setActivePanel("file");
@@ -2495,20 +408,9 @@ export default function WorkspaceTreeView() {
             expandedItems={expandedItems}
             selectedItems={selectedItems}
             renderers={{
-              renderTreeContainer: (props) => {
-                return (
-                  <>
-                    <ul {...props.containerProps}>{props.children}</ul>
-                  </>
-                );
-              },
-              renderItemsContainer: (props) => {
-                return (
-                  <>
-                    <ul {...props.containerProps}>{props.children}</ul>
-                  </>
-                );
-              },
+              renderTreeContainer: (props) => <Container {...props} />,
+              renderItemsContainer: (props) => <Container {...props} />,
+
               renderItem: (props) => {
                 useEffect(() => {
                   const node = props.item.data as TNode;
@@ -2517,6 +419,7 @@ export default function WorkspaceTreeView() {
                     props.context.startRenamingItem();
                   }
                 }, []);
+
                 const fileReferenceData = useMemo<TFilesReference>(() => {
                   const node = props.item.data as TNode;
                   const nodeData = node.data as TFileNodeData;
@@ -2597,6 +500,10 @@ export default function WorkspaceTreeView() {
                   }
                 }
 
+                const onMouseEnter = () => setFFHoveredItem(props.item.index as TNodeUid)
+
+                const onMouseLeave = () => setFFHoveredItem("" as TNodeUid)
+                
                 return (
                   <>
                     <li
@@ -2639,10 +546,8 @@ export default function WorkspaceTreeView() {
                         {...props.context.interactiveElementProps}
                         onClick={onClick}
                         onFocus={() => {}}
-                        onMouseEnter={() =>
-                          setFFHoveredItem(props.item.index as TNodeUid)
-                        }
-                        onMouseLeave={() => setFFHoveredItem("" as TNodeUid)}
+                        onMouseEnter={onMouseEnter}
+                        onMouseLeave={onMouseLeave}
                         onDragStart={onDragStart}
                         onDragEnter={onDragEnter}
                       >
@@ -2691,21 +596,7 @@ export default function WorkspaceTreeView() {
                   </>
                 );
               },
-              renderItemArrow: (props) => {
-                return (
-                  <>
-                    {props.item.isFolder ? (
-                      props.context.isExpanded ? (
-                        <SVGIconI {...{ class: "icon-xs" }}>down</SVGIconI>
-                      ) : (
-                        <SVGIconII {...{ class: "icon-xs" }}>right</SVGIconII>
-                      )
-                    ) : (
-                      <div className="icon-xs"></div>
-                    )}
-                  </>
-                );
-              },
+              renderItemArrow: (props) => <ItemArrow {...props}/>,
               renderItemTitle: (props) => {
                 return (
                   <>
@@ -2733,6 +624,26 @@ export default function WorkspaceTreeView() {
                 );
               },
               renderRenameInput: (props) => {
+
+                const onChange = useCallback( 
+                    (e: React.ChangeEvent<HTMLInputElement>) => {
+                    props.inputProps.onChange &&
+                      props.inputProps.onChange(e);
+                  },[props.inputProps]
+                )
+
+                const onBlur = useCallback(
+                  (e: React.FocusEvent<HTMLInputElement, Element>) => {
+                    props.inputProps.onBlur && props.inputProps.onBlur(e);
+                    props.formProps.onSubmit &&
+                      props.formProps.onSubmit(
+                        new Event(
+                          "",
+                        ) as unknown as React.FormEvent<HTMLFormElement>,
+                      );
+                  },[props.inputProps, props.formProps]
+                )
+
                 return (
                   <>
                     <form {...props.formProps} className={"box"}>
@@ -2748,19 +659,8 @@ export default function WorkspaceTreeView() {
                           padding: "0",
                           background: "transparent",
                         }}
-                        onChange={(e) => {
-                          props.inputProps.onChange &&
-                            props.inputProps.onChange(e);
-                        }}
-                        onBlur={(e) => {
-                          props.inputProps.onBlur && props.inputProps.onBlur(e);
-                          props.formProps.onSubmit &&
-                            props.formProps.onSubmit(
-                              new Event(
-                                "",
-                              ) as unknown as React.FormEvent<HTMLFormElement>,
-                            );
-                        }}
+                        onChange={onChange}
+                        onBlur={onBlur}
                       />
                       <button
                         ref={props.submitButtonRef}
