+++ src/components/main/actionsPanel/workspaceTreeView/workSpaceTreeComponents/index.ts
@@ -1,0 +1,2 @@
+export * from './Container';
+export * from './ItemArrow';
